refactor: code-editor
