+++ electron/src/frontend/panels/editor/components/code-editor.tsx
